Move basic elements to own file
